Fix: 네트워크에 있어야하는 부분 분리
