--- ORI/Utils/Network/LoginNetwork.swift
@@ -1,7 +1,0 @@
-//
-//  LoginNetwork.swift
-//  ORI
-//
-//  Created by Song Kim on 3/5/25.
-//
-
