+++ ORI/View/LoginViewController.swift
@@ -58,7 +58,7 @@ class LoginViewController: UIViewController {
     @objc func handleGoogleLogin() {
         viewModel.signInWithGoogle(presentingViewController: self) { success in
             if success {
-                self.viewModel.requestTokenFromServer(idToken: self.viewModel.idToken)
+                TokenNetwork.requestTokenFromServer(idToken: self.viewModel.idToken)
                 NavigationManager.navigateToTabView()
             } else {
                 print("Google login failed")
